# backend/api/swagger.yaml
swagger: "2.0"
info:
  title: Accountability Expense-Tracker API
  version: 1-oas3
host: accountability-190955e8b06f.herokuapp.com
basePath: /api
schemes:
  - https
paths:
...
  /expenses/create:
    post:
      summary: CREATE expense endpoint
      description: Create a new, blank expense
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            tripId:
              type: string
              description: ID of the trip this expense will belong to
            name:
              type: string
            description:
              type: string
            cost:
              type: number
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              expenseId:
                description: ID of the newly created expense
                type: string
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /expenses/get:
    post:
      summary: READ expense endpoint
      description: Read information about an expense. Note that /trips/get can return all the expenses at once.
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            expenseId:
              type: string
              description: ID of the expense to delete
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              _id:
                type: string
              name:
                type: string
              description:
                type: string
              cost:
                type: number
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /expenses/update:
    post:
      summary: UPDATE expense endpoint
      description: Modify expense properties
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            expenseId:
              type: string
              description: ID of the expense to modify
            name:
              type: string
            description:
              type: string
            cost:
              type: number
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              expenseId:
                description: Same as input ID
                type: string
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /expenses/delete:
    post:
      summary: DELETE expense endpoint
      description: Remove an expense and associated data
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            expenseId:
              type: string
              description: ID of the expense to delete
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'      
definitions:
...
    ErrorInfo:
      type: object
      properties:
        error:
          type: string
          description: The text description of the error.
          example: Information about the error
responses:
  BadRequest:
    description: Malformed request or JSON on front-end
    schema:
      $ref: '#/definitions/ErrorInfo'
...